Fixed some warnings
git-svn-id: http://hcm-lab.de/public/Horde3D/trunk@203 5ce291ac-9df0-446f-9e4f-d57731c4dda7

+++ Extensions/Sound/Source/decode_vorbis.cpp
@@ -63,14 +63,14 @@ namespace Horde3DSound
 			if( offset > vorbisData->size )
 				vorbisData->sizeRead = vorbisData->size;
 			else
-				vorbisData->sizeRead = offset;
+				vorbisData->sizeRead = (size_t)offset;
 
 			break;
 		case SEEK_CUR:
 			if( offset > vorbisData->size - vorbisData->sizeRead )
 				vorbisData->sizeRead = vorbisData->size;
 			else
-				vorbisData->sizeRead += offset;
+				vorbisData->sizeRead += (size_t)offset;
 
 			break;
 		case SEEK_END:
